Fallback to old chem checking if new method fails
While the old checking mechanism treated the \ce macro
as a regular math command with one argument the argument
was rewritten according to texvc spacing rules.

This caused extra spaces, which were interpreted by the
mhchem logic. To avoid these unwanted spaces people massaged
the input in a way that it looked good in the end by adding
extra spacing information. However, the final result is not
always compatible with the new ce grammar rules.

While there might be different classes of errors manual investigtion
is required for some cases. To allow for editors to review the problems
a new warning mechanism is introduced.

The response has now an array with warnings.
The only warning currently produced is the warning that
indicates that \ce commands are used that were evaluated
against the old checking mechanism.

cf. https://phabricator.wikimedia.org/T183557